[Docs] Add limitation about copying target dir files to remote (#1305)
It’s currently observed that users may assume, after reading the
[documentation](https://astronomer.github.io/astronomer-cosmos/configuration/cosmos-conf.html#remote-target-path),
that all files from the target directory of the dbt project are uploaded
to the remote target directory. However, this isn’t yet the case. The
`remote_target_path` configuration, introduced in PR
https://github.com/astronomer/astronomer-cosmos/pull/1224, currently
only uploads files from the compiled directory within the target
directory of the dbt project—and only when the execution mode is set to
`ExecutionMode.AIRFLOW_ASYNC`. This PR updates the documentation for
this configuration to clarify this behavior and indicate that additional
files will become available as we expand support for uploading them
within the operators.

closes: #1291