Add postgres pod status check for k8s tests in CI (#1320)
It appears we have a flaky Kubernetes test that failed in PR #1313. As
shown in the error log
[here](https://github.com/astronomer/astronomer-cosmos/actions/runs/11796817624/job/32867560902?pr=1313#step:7:473),
the PostgreSQL pod did not reach the ready state and instead entered an
error status. Since the cause of the error status is unclear, this PR
introduces a status check for the PostgreSQL pod to ensure it becomes
fully running and healthy. If the pod enters an ERROR state, we now run
a `kubectl describe` command on the pod to capture the event logs for
debugging. The test will also exit with an error code of 1 to prevent
further execution.

related: #1319