Add exempt label option in GH action stale job (#1328)
I’ve noticed that the "dosubot" and "stale" jobs are creating
unnecessary noise by closing issues and PRs that were not intended to be
closed. In this PR, I am introducing the exempt-issue-labels and
exempt-pr-labels options for the stale job. These options will ensure
that issues and PRs labeled with do-not-stale are excluded from being
closed by the "dosubot" and "stale" jobs.