Enable Azure task in the remote store manifest example DAG (#1333)
During the work on PR
https://github.com/astronomer/astronomer-cosmos/pull/1297, an issue
arose where the Azure remote manifest task began failing after
installing providers and packages with constraints. To allow other
tests, which were running successfully, to proceed, the task was
temporarily disabled. Upon reviewing the GitHub Actions logs from
previous successful runs, it was
[observed](https://github.com/astronomer/astronomer-cosmos/actions/runs/11573670971/job/32216315282#step:6:250)
that the Azure provider version installed was 10.5.1. However, after the
refactoring introduced in the PR, the failing actions
[showed](https://github.com/astronomer/astronomer-cosmos/actions/runs/11911545582/job/33193301710#step:6:474)
azure provider==8.4.0 being installed with the constraints file.

To investigate, I tested locally to identify a working version. While
the task failed with 8.4.0, it succeeded with 8.5.0. Analyzing the
failure
[logs](https://github.com/astronomer/astronomer-cosmos/actions/runs/11911545582/job/33193301710#step:7:467)
and reviewing the Azure provider changelog hints that
https://github.com/apache/airflow/pull/35820/ is potentially the fix for
the failure with our connection setup in the CI that was released in
8.5.0.

Therefore, I propose using azure provider>=8.5.0. However, the Airflow
[constraints
file](https://raw.githubusercontent.com/apache/airflow/constraints-2.8.0/constraints-3.8.txt)
for version 2.8 specifies azure provider==8.4.0, which conflicts with
this requirement. To address this, I am making changes to the
pre-install script in our CI to install azure provider>=8.5.0 without
relying on the constraints file, citing the reasons above.

closes: #1304