Mention in documentation that the callback functionality is supported in ExecutionMode.VIRTUALENV (#1401)
Following up on PR #1389 wherein we add helper functions to demonstrate
how to
utilise the callback functionality in `ExecutionMode.LOCAL`, I also
tested this functionality
with `ExecutionMode.VIRTUALENV` since the VirtualEnv mode operators
inherit the Local
mode operators and it does seem to support the callback functionality
well. I was well able
to test this by adding the callback arg to the VirtualEnv example DAG we
have and see that
the files do get uploaded to remote store. I have added the argument in
comments in the example
DAG so that users could take a reference of that, however I have left
the argument as commented
so that we do not always keep on uploading the files in our CI runs and
given that we already this
in the `example_operators.py` DAG.

The PR also updates the documentation to reflect that the callback
functionality is supported in
the `ExecutionMode.VIRTUALENV` too.

closes: #1399