Fix Snowflake Profile mapping when using AWS default region (#1406)
When using a Cosmos Snowflake Profile mapping using a Snowflake account
set in the AWS default region, Cosmos would fail if the default region
was specified in the Airflow connection.

The dbt docs state:
> For AWS accounts in the US West default region, you can use abc123
(without any other segments). For some AWS accounts you will have to
append the region and/or cloud platform. For example, abc123.eu-west-1
or abc123.eu-west-2.aws.

https://docs.getdbt.com/docs/core/connect-data-platform/snowflake-setup#account
 
Although it seems that defining the default region would be optional, a
Cosmos user reported facing 404 and seeing a dbt error message when
attempting to use `SnowflakeUserPasswordProfileMapping` with an Airflow
Snowflake connection that defined the region `us-west-2`.

![snowflake-404](https://github.com/user-attachments/assets/c1884fff-1cad-4c57-b2f3-11a4f44b085b)

We solved the issue by removing the region `us-west-2` from the
connection.

Since this restriction only applies to AWS and this Snowflake region
only exists to AWS, this change seems safe:
![Screenshot 2024-12-18 at 18 45
31](https://github.com/user-attachments/assets/ff2f8a0b-578b-4a62-9fc3-258a43148775)