Drop running 1000 models in the CI (#1411)
With PR #1403 , we added to the matrix to run tests on 500 & 1000
models. Although the job for
500 models completes in under 15 mins, the job for 1000 models is taking
significantly longer ~40+ mins as can be seen here
https://github.com/astronomer/astronomer-cosmos/actions/runs/12400943145/job/34619249671.

I suggest we drop running 1000 models and later add it to the matrix to
conditionally run only when a certain label is added to the PR i.e. when
we would actually like to benchmark on specific PRs. I have created a
follow-up issue for this
https://github.com/astronomer/astronomer-cosmos/issues/1410

related: #1410