Enable Docs DAG in CI leveraging existing CI connections (#1428)
It seems that the dbt docs DAG in our example DAGs suite was silently
failing to run because the specified connections in the DAG were not
present in our CI environment. I am removing the connection checks,
allowing the previously skipped tasks to execute even when the remote
Airflow connections are unavailable. Additionally, I’ve updated the DAG
to use our existing CI credentials and Airflow connections, as well as
updated the bucket name to the ones I created in our object stores. This
change will help catch issues like #1420 earlier. While fix #1422 has
already addressed #1420, this PR will now validate that the fix is
functioning as expected.

closes: #1420