Allow users to opt-in or out (default) of detached test nodes (#1470)
Since we introduced detached test tasks to fix a customer issue in #1433
(release 1.8.1), we changed how Cosmos renders DAGs, with the chance
that Cosmos significantly changed how it renders a dbt project - even
when users did not change their `DbtDag` or `DbtTaskGroup`
configuration. This is unacceptable in a micro release - and for this
reason we're reverting this change and making the feature opt-in.

PR #1433 led to issues such as #1464, reported by multiple Cosmos users,
and also issues that did not become Github issues, such as an Astro
customer who reported that when they upgraded to Cosmos 1.8.1, the
number of tasks increased dramatically. One problem with #1433 was that
it did not empower users to opt in or out of having detached test nodes,
solving the problem for some but causing problems for many.

This PR aims to solve this problem by introducing a new property to
`RenderConfig`: `should_detach_multiple_parents_tests`. We are reverting
the Cosmos DAG rendering to what it was in 1.8.0 and before: by default,
it will not detach tests with multiple parents. Users must opt-in for
this behaviour if and when they want to.

We understand this may be perceived as a breaking change by some, but it
is the correct way to move forward and avoid causing further disruption.

We are planning to review the current implementation, as described in
#1469. For now, this PR documents the current behaviour and
empowers users.