Fix import error in dbt bigquery adapter mock for `dbt-bigquery<1.8` for `ExecutionMode.AIRFLOW_ASYNC` (#1548)
A user has reported after testing the `astronomer-cosmos==1.9.0a5` that
they are getting the below error
```
    from dbt_common.clients.agate_helper import empty_table
ModuleNotFoundError: No module named 'dbt_common'
```

They are using `dbt-bigquery==1.7.2`

Upon debugging, I observed that the `dbt_common` module that we rely on
in the current mocking interface is available only in dbt bigquery
adapterversion >= 1.8. For previous versions, to achieve the same, the
helper seems to be available in `dbt.clients`. I tested this on
dbt-bigquery versions 1.5, 1.6, 1.7 and 1.7.2 and the fix in this PR
seems to solve the issue.

closes: #1547