Support `on_warning_callback` with `TestBehavior.BUILD` and `ExecutionMode.LOCAL` (#1571)
As of now, when we set TestBehavior.BUILD, we are not leveraging the
method on_warning_callback that is available for Test nodes and Source
Nodes. I have added the parsing to DbtBuildLocalOperator in order to fix
it. I tested it locally and I got positive results

Related: https://github.com/astronomer/astronomer-cosmos/issues/1569