Users who generated the `manifest.json` using MS Windows and attempted to use Cosmos path selectors after, such as `path:models/edr/run_results' were unable to do so, because the paths in Windows were different from the selector:
```
    "model.elementary.model_run_results": {
      "database": "FDH_DEV_DB",
      "schema": "MONITORING",
      "name": "model_run_results",
      "resource_type": "model",
      "package_name": "elementary",
      "path": "edr\\run_results\\model_run_results.sql",
      "original_file_path": "models\\edr\\run_results\\model_run_results.sql",
      "unique_id": "model.elementary.model_run_results",
      "fqn": [
        "elementary",
        "edr",
        "run_results",
        "model_run_results"
      ],
```

As it can be observedf in this example, the property `original_file_path` used the `\\` character as a divider in the path,
but the selector checked using the Posix notation.

Since Cosmos implements path selectors using: path_selection in str(node.file_path), we have to normalize the input for the
filter to work.

This issue only happened when using `LoadMode.DBT_MANIFEST` and not `LoadMode.DBT_LS`, since dbt normalizes this internally
when handling selectors as part of this command line.