Support filtering by config meta nested properties (#1617)
Allow users to use `select` and `exclude` based on `config.meta` when
using `LoadMode.MANIFEST`.

An Astro Customer request reported in:
https://astronomer.slack.com/archives/C027CNLBGG7/p1741967575692289?thread_ts=1741881283.263679&cid=C027CNLBGG7.

> For the selector issue, it has to do with the model tagging I
mentioned above. Right now, with tags, if a user set the following tags
in the `dbt_project.yml`:
> ```
> models:
>   +tags=["core_models", "daily"]
> ```
> but they have a model under models/ that needs to run monthly, so they
set the following in the model config block:
> ```
> {{
>   config(
>    tags=["monthly"]
>  )
>}}
>```
> That specific model ends up with the following tags: ["core_models",
"daily", "monthly"] . This is because tags are additive when applied
this way. With the way our Cosmos configs are setup (daily dag provided
as an example):
>```render_config=RenderConfig(
>  load_method=LoadMode.DBT_MANIFEST,
>  select=["tag:core_models", "tag:daily"],
>  exclude=[
>    "tag:hourly",
>    "tag:weekly",
>    "tag:monthly",
>    "tag:static",
>    "tag:analyst_scheduled",
>  ],
>)
> ```
> This would prevent the task that's supposed to run monthly from ever
being scheduled.
> Since selects are additive, we can't just have:
>
> ```render_config=RenderConfig(
>  load_method=LoadMode.DBT_MANIFEST,
>  select=["tag:core_models", "tag:daily"],
>)
>```
> as this would pull all models with either core_models or daily tags,
which isn't our desired goal.
> In order to resolve this, we are discussing avenues to define the
dbt_environment and frequency in a more unique way. Our current thought
process is to utilize the meta section of the model config and set the
following for models:
>
>```models:
>  <model-dir>:
>    +meta: {
>      "dbt_environment": "<dbt_environment>",
>      "frequency": "<frequency>"
>    }
>```
>This way, we can define defaults in the dbt_project.yml, but then
override at a lower granularity when need be. However, in order for this
to work with cosmos, we need the ability to select models based on this
config, something like config:meta.dbt_environment and
config:meta.frequency. Based on our limited testing, this didn't seem
possible with `DBT_MANIFEST` load method, but did seem possible with
`DBT_LS`

Closes: #1619