docs: Add MonteCarlo callback example for importing dbt artifacts (#1695)
Add an extra example on how to import dbt artifacts using montecarlo
integration with dbt-core as a callback. The implementation is quite
challenging considering the lack of documentation provided by montecarlo
itself for the python SDK and after struggling to implement something
functional, I believe it is worth having it in the cosmos documentation
since the line above the code-block mentions montecarlo, suggesting that
many users would be interested in that.