Use latest minor versions for dbt adapters to get in compatibility fixes (#1719)
While the initial x.x.0 versions of dbt adapters may have compatibility
issues with certain dependencies, I observed that these
incompatibilities are progressively addressed in later patch releases.
The issue reported in #1709 appears to stem from dbt-bigquery accessing
a protected member `google.cloud.bigquery._helpers._CELLDATA_FROM_JSON`,
which was removed in `google-cloud-bigquery 3.31.0`. This access I
believe was introduced in
[dbt-bigquery#974](https://github.com/dbt-labs/dbt-bigquery/pull/974)
and later removed in
[dbt-bigquery#1061](https://github.com/dbt-labs/dbt-bigquery/pull/1061),
with the fix subsequently back-ported to the 1.7 series via
[dbt-bigquery#1074](https://github.com/dbt-labs/dbt-bigquery/pull/1074).
Therefore, I believe that relying on the latest patch versions is a
better approach to avoid such issues, rather than individually resolving
discrepancies. This PR aligns with that strategy.


closes: #1709

---------

Co-authored-by: Tatiana Al-Chueyr <tatiana.alchueyr@gmail.com>