Operator argument fix (#1648)
Cosmos 1.9.0 introduced a change such that `AbstractDbtBase` no longer
inherits from airflow `BaseOperator` and I believe this introduced two
bugs.

- In the past you could give arguments using airflows `default_args` and
these would end up being passed to cosmos. So for example we were
specifying the `project_dir` as a default arg, and in 1.9.0 these
projects were broken because AbstractDbtBase was not longer getting a
project_dir specified. So the first change is to check both kwargs and
default_args for the values AbstractDbtBase requires. ~I've only made
this change for the kubernetes operator, but I'd argue it should be done
for the local a docker ones too.~
- The existing code tries to inspect `KubernetesPodOperator` for the
args that it requires, but notably this approach misses arguments that
are required by `BaseOperator`, e.g. the task_group.
- ~I've made an effort to restore the convention in airflow that all
arguments must be consumed by the operator. So an argument must either
be consumed by `AbstractDbtBase` or it will be passed along to the
operator, it may also be consumed by both classes.~