Allow multiple callbacks (#1693)
During some experimentation with
[callbacks](https://astronomer.github.io/astronomer-cosmos/configuration/callbacks.html#example-using-callbacks-with-a-single-operator)
I have noticed that we can only use a single callback per time, which
can be very limiting if I want to add custom callbacks together with the
default ones provided by cosmos.

This PR is very simple and aims to allow multiple callbacks to be passed
as a list of functions, that will then be executed sequentially.