Enable example DAGs to run in CI that were disabled in PR #1646 (#1726)
Airflow 3.0 does not currently support calling `.test()` on DAGs defined
using the `@dag` decorator, as the decorator now returns an object that
does not expose the `.test()` method. There is an open issue
https://github.com/apache/airflow/issues/45549 around this area. This PR
updates the affected example DAGs to explicitly use `airflow.models.DAG`
instead of the `@dag` decorator, restoring compatibility with
`dag.test()` for local testing and CI.

This also re-enables the example DAGs that were disabled in PR #1646.

closes: #1706