Run listener tests in Airflow 3 (#1743)
closes: https://github.com/astronomer/astronomer-cosmos/issues/1703

Previously, we sent the dag_hash to Scarf,
but with the removal of DagRun.dag_hash in Airflow 3,
this PR modifies the implementation to send a portion
of the dag_id hash instead. This change is a reasonable
compromise, as we only need a unique identifier for the DAG.
Additionally, the PR updates the relevant test cases to reflect this
change.

---------

Co-authored-by: Tatiana Al-Chueyr <tatiana.alchueyr@gmail.com>
(cherry picked from commit a532a20376615d435d0f6f793a97b1f2541d3781)