Fix hash value for Darwin when using Py 3.12.10
In the past, we saw inconsistencies in how hashing worked depending on the platform
(Darwin x Linux). When setting up Cosmos in a brand new laptop that runs MacOS 15.4.1 (Arm 64)
I noticed the hash was the same as the CI.

In order to not break for others, I'm still supporting the other value for MacOS.

This is a test-only change and does not affect the actual implementation of Cosmos.