Allow to set task display name as a user-defined function (#1761)
Re-opened #1756 after resetting my fork by mistake.

> Currently cosmos allows to override the task_id value with the param
[normalize_task_id](https://astronomer.github.io/astronomer-cosmos/configuration/task-display-name.html),
but there is no process in place to override the display_name . At my
current company we have some logic in place using the task_id and
corresponding suffixes, but we would like to still change how tasks are
displayed in the airflow UI, plus from a user's perspective it is
sometimes confusing to have the suffix being displayed in the airflow UI
instead of just having the name of the model itself. It would be nice to
have a way to keep the task_id as it is while overriding the
display_name instead so if a user is already using the task_ids for some
custom logic, it is possible to change how the tasks are displayed in
the UI without causing issues.

> a small disclaimer: I took the liberty to rename task_display_name to
task_name in the code since it was not exactly correct, and since this
PR aims to separate both concepts, task_name is a more generic name in
case that the display_name is not set 😄

Co-authored-by: GiovanniCorsetti <66136602+CorsettiS@users.noreply.github.com>