fix: Convert Kubernetes Pod Operator `container_resources` (#1821)
## Description

Slightly odd and I'm not sure if this is the best solution, but this is
the code base I have access to, so lets go with it.

I'm trying to specify resource limits in cosmos using
https://astronomer.github.io/astronomer-cosmos/configuration/operator-args.html#overriding-operator-arguments-per-dbt-node-or-group-of-nodes,
specifically I'm adding
```yaml
    meta:
      cosmos:
        operator_kwargs:
          container_resources:
            limits:
              memory: "2000Mi"
```
to a DBT project and when running locally using `astro dev start` this
works correctly. However when running in astronomer hosted, it doesn't
apply. Researching further if I have an operator with
```
KubernetesPodOperator(
        ...
        container_resources={
            "limits": {"cpu": "100m", "memory": "100Mi", "ephemeral-storage": "512Mi"},
            "requests": {
                "cpu": "100m",
                "memory": "100Mi",
                "ephemeral-storage": "512Mi",
            },
        },
    )
```
then the resources will be overridden with the defaults set on the
deployment in astronomer hosted, though running locally it works
correctly. If I specify them as
```
KubernetesPodOperator(
        ...
        container_resources=k8s.V1ResourceRequirements(
            limits={"cpu": "100m", "memory": "100Mi", "ephemeral-storage": "512Mi"},
            requests={"cpu": "100m", "memory": "100Mi", "ephemeral-storage": "512Mi"},
        ),
    )
```
then it all works correctly, but this precludes my ability to set them
in cosmos as the argument will always be passed as a dictionary there.

It seems whatever code astronomer is using on the backend to apply the
default container resources doesn't work with anything except
`V1ResourceRequirements`. Arguably this is correct behaviour since the
signature in the kubernetes pod operator specifies the type to be
`V1ResourceRequirements | None`, even though the code itself does work
with dictionaries.

Anyway the proposed solution is to convert this argument to the correct
type before it is passed to the kubernetes pod operator. Thoughts?

---------

Co-authored-by: Tatiana Al-Chueyr <tatiana.alchueyr@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>