CI: Add back accidentally deleted python-version matrix for running unit tests (#1872)
Looks like the CI is failing to run tests after having accidentally
dropped the matrix entries for python-version in
https://github.com/astronomer/astronomer-cosmos/pull/1807

This came to my attention while I'm working on another PR whose action
run complained about this in
https://github.com/astronomer/astronomer-cosmos/actions/runs/16320855168

We are getting a yml validation error for the missing matrix key. The PR
fixes this issue by adding back the deleted matrix entry.

---------

Co-authored-by: Tatiana Al-Chueyr <tatiana.alchueyr@gmail.com>