Fix cache behaviour when DAG name contains .
Cosmos stores the result of dbt ls as a Variable based on the DAG name, but if the name has a . in it, that Variable name will be invalid when working with certain secret managers, including Google cloud's.

This is an example of error that is raised:
```
Invalid secret ID myhost-cosmos_cache__dag.moredag.yetmoredag
Only ASCII alphabets (a-Z), numbers (0-9), dashes (-), and underscores (_) are allowed in the secret ID.
```

In this PR, we replace those names to use underscores instead of dot.

Closes: #1875