Push producer's task completion status to XCOM (#2000)
This PR pushes a `task_status` XCOM for the `DbtProducerWatcherOperator`
task after its completion -- be it failed or succeeded.

The idea is that sensors will pull this XCOM while checking for the
result of the model. If a sensor sees that the
`DbtProducerWatcherOperator` task has completed (either succeeded or
failed), but the `DbtProducerWatcherOperator` has not pushed the model
status to XCOM—whether via the `node_finished...` XCOM or the
`run_results` XCOM—it means the operator has already finished processing
and will not push any further status updates.

In that case, the sensor should immediately mark the model as failed,
since there is no chance of receiving a run status for it in the future,
and the sensor task itself should also fail.

Related: https://github.com/astronomer/oss-integrations-private/issues/239
Related: https://github.com/astronomer/astronomer-cosmos/issues/1960