Update setup script for airflow3 script (#2023)
I'm setting up Airflow 3 for the first time on my laptop and ran into
some issues with the setup as-written. I implemented some relatively
conservative changes to the development environment setup that is within
`scripts/airflow3`.

- `env.sh` script does not have execute permission (644); changed to
755.
- "set +e" is required at the end of the env.sh script to prevent it
from closing out.
- - In requirements.txt: psycopg2 -> psycopg2-binary. Via psycopg2
readme: "The binary package is a practical choice for development and
testing but in production it is advised to use the package built from
sources." Since these is a development environment, it should be
preferred and reduces issues.
- I don't believe asyncpg is necessary; the psycopg2 driver is
completely independent of asyncpg.