Fix task flow operator args (#2024)
Similar to the issue
[here](https://github.com/astronomer/astronomer-cosmos/pull/1360). With
task flow for operator_args, copy.deepcopy(operator_args) was raising an
error due to a completely new copy of the taskflow output being passed
in. Using copy.copy(operator_args) instead means we're still referencing
the original taskflow output, but still have a new copy of the
operator_args itself.

Note: I was struggling to get the cosmos tests to run locally so
apologies that I haven't actually run the new test! But I did recreate
the issue in my local airflow environment, install this version of the
package, and confirm that this seems to fix it.

Closes: #1984