Fix `ExecutionMode.WATCHER` behaviour with `DbtTaskGroup` (#2044)
While validating `TaskGroup` in a session with an
[end-user](https://us-35700.app.gong.io/call?id=1548801879239330851&email_type=call-ready-notification&xtid=45fp79m8frjifyiyqp0)
with Cosmos
[v1.11.0a8](https://github.com/astronomer/astronomer-cosmos/releases/tag/astronomer-cosmos-v1.11.0a8),
we identified that it was not working as expected.

There were two main issues:
1. We were not setting the `producer_task_id` argument correctly in the
sensor tasks
2. The producer task was always being picked up after the sensor tasks,
even though it had a higher priority weight and was an "upstream" task
of the dbt root nodes (with a trigger rule always, so they could start
sensing).

The first issue was a bug, which is now fixed.

The second issue occurs because the sensor tasks have a `trigger_rule`
of "always". This setting works well in the case of `DbtDag`, as the
only upstream task is the producer task, and the producer task and root
dbt project nodes can be scheduled at the same time, with the task
weight of the producer giving its precedence. However, in the case of a
`DbtTaskGroup` with one or multiple upstream tasks, the producer task
must wait for the upstream task, whereas the dbt project root nodes
("top sensors") do not have to wait. This leads to sensors occupying the
available lots before the producer is scheduled to arrive.

At some point, we suspected this was an issue with Airflow itself, but
after a troubleshooting session with Ash, we realised it was not a bug
in Airflow: https://github.com/apache/airflow/issues/56723.

The side-effect of the implementation in this PR is that `DbtTaskGroup`
and `DbtDag` are behaving differently:
- In `DbtDag`, the producer is upstream of the dbt root nodes
- In `DbtTaskGroup`, the producer is not an upstream root node

The reason we are keeping this behaviour in `DbtDag` is that it is not
leading to any issues. It is topologically "correct", making it easier
for users to see the dependency between the tasks & potentially view the
task list topologically sorted.

At this moment, we were unable to find a way to offer this same
behaviour with `DbtTaskGroup`. @ashb , with whom I troubleshooted the
original "problem", pointed out Airflow could potentially have a
different "state" that wouldn't always be, but something like "start
when my upstream task started". However, by examining [Airflow's current
implementation](https://github.com/apache/airflow/blob/c160ab70a00faa41ef5996c4ba4a6580a64194ad/airflow-core/src/airflow/ti_deps/deps/trigger_rule_dep.py),
any changes to this area could potentially incur an additional cost to
the scheduler and compromise Airflow's performance.

If we were able to "force" in the Airflow UI to have the producer task
on the "top" both in the graph view and in the list of functions view,
we could make `DbtDag`and `DbtTaskGroup` more consistent (by not having
the producer task as upstream and dbt root nodes with trigger_rule
always). Currently, in Airflow 3.0.x and Airflow 3.1.0, the producer
task has been showing randomly, unfortunately, but we will work with the
Airflow team to improve this and add follow-up tickets as needed.

Closes: https://github.com/astronomer/astronomer-cosmos/issues/1961
Closes:
https://github.com/astronomer/oss-integrations-private/issues/240